wrong long cast in fromReference

+++ src/IHM/model/SI.java
@@ -68,6 +68,7 @@ public abstract class SI extends Unit
 		this.pas = pas;
 		this.prefix = Math.pow(10, (SI.PREFIX.get(prefix.replace(base, "")) * this.pas));
 		this.prefix_ref = Math.pow(10, (SI.PREFIX.get(prefix_ref.replace(base, "")) * this.pas));
+		//System.out.println(this.prefix+ " " + this.prefix_ref);
 	}
 	
 	public void setReference(String reference){
@@ -107,7 +108,7 @@ public abstract class SI extends Unit
 	 */
 	public void fromReference(double qtOfRef)
 	{
-		this.setQuantity( (long) ((qtOfRef / this.prefix) * this.prefix_ref) );
+		this.setQuantity( ((qtOfRef / this.prefix) * this.prefix_ref) );
 	}
 
 	/**
